refactor: consolidate Convex functions to eliminate circular dependencies
- Move all queries, mutations, and actions to root messages.ts (1211 lines)
- Convert messages/ directory to pure utility functions with no database calls
- Remove empty placeholder files (queries.ts, mutations.ts, actions.ts, conversation.ts)
- Fix TypeScript circular type inference "deep instantiation" errors
- Optimize imports and remove unused dependencies
- Maintain clean separation: database ops in root, utilities in messages/

This architectural change completely resolves circular dependency issues
while improving code organization and maintainability.
