+++ apps/www/convex/_generated/api.d.ts
@@ -21,11 +21,11 @@ import type * as http from "../http.js";
 import type * as lib_ai_client from "../lib/ai_client.js";
 import type * as lib_ai_tools from "../lib/ai_tools.js";
 import type * as lib_auth from "../lib/auth.js";
+import type * as lib_conversation from "../lib/conversation.js";
 import type * as lib_database from "../lib/database.js";
 import type * as lib_encryption from "../lib/encryption.js";
 import type * as lib_errors from "../lib/errors.js";
 import type * as lib_message_builder from "../lib/message_builder.js";
-import type * as lib_message_service from "../lib/message_service.js";
 import type * as messages from "../messages.js";
 import type * as messages_actions from "../messages/actions.js";
 import type * as messages_helpers from "../messages/helpers.js";
@@ -58,11 +58,11 @@ declare const fullApi: ApiFromModules<{
 	"lib/ai_client": typeof lib_ai_client;
 	"lib/ai_tools": typeof lib_ai_tools;
 	"lib/auth": typeof lib_auth;
+	"lib/conversation": typeof lib_conversation;
 	"lib/database": typeof lib_database;
 	"lib/encryption": typeof lib_encryption;
 	"lib/errors": typeof lib_errors;
 	"lib/message_builder": typeof lib_message_builder;
-	"lib/message_service": typeof lib_message_service;
 	"messages/actions": typeof messages_actions;
 	"messages/helpers": typeof messages_helpers;
 	"messages/mutations": typeof messages_mutations;
