+++ apps/www/convex/lib/message_builder.ts
@@ -1,96 +1,4 @@
 import { type ModelId, getModelById } from "../../src/lib/ai/schemas.js";
-import { internal } from "../_generated/api.js";
-import type { Id } from "../_generated/dataModel.js";
-import type { ActionCtx } from "../_generated/server.js";
-
-// Type definitions for multimodal content based on AI SDK v5
-type TextPart = { type: "text"; text: string };
-type ImagePart = { type: "image"; image: string | URL };
-type FilePart = {
-	type: "file";
-	data: string | URL;
-	mediaType: string;
-};
-
-export type MultimodalContent = string | Array<TextPart | ImagePart | FilePart>;
-
-/**
- * Build message content with attachments
- * Single source of truth for content building logic
- */
-export async function buildMessageContent(
-	ctx: ActionCtx,
-	text: string,
-	attachmentIds?: Id<"files">[],
-	provider?: "openai" | "anthropic" | "openrouter",
-	modelId?: ModelId,
-): Promise<MultimodalContent> {
-	// If no attachments, return simple text content
-	if (!attachmentIds || attachmentIds.length === 0) {
-		return text;
-	}
-
-	// Get model configuration to check capabilities
-	const modelConfig = modelId ? getModelById(modelId) : null;
-	const hasVisionSupport = modelConfig?.features.vision ?? false;
-	const hasPdfSupport = modelConfig?.features.pdfSupport ?? false;
-
-	// Build content array with text and files
-	const content = [{ type: "text" as const, text }] as Array<
-		TextPart | ImagePart | FilePart
-	>;
-
-	// Fetch each file with its URL
-	for (const fileId of attachmentIds) {
-		// @ts-ignore - TypeScript deep instantiation issue - using @ts-ignore for now
-		const file = await ctx.runQuery(internal.files.getFileWithUrl, { fileId });
-		if (!file || !file.url) continue;
-
-		// Handle images
-		if (file.fileType.startsWith("image/")) {
-			if (!hasVisionSupport) {
-				// Model doesn't support vision
-				if (content[0] && "text" in content[0]) {
-					content[0].text += `\n\n[Attached image: ${file.fileName}]\n⚠️ Note: ${modelConfig?.displayName || "This model"} cannot view images. Please switch to GPT-4o, GPT-4o Mini, or any Claude model to analyze this image.`;
-				}
-			} else {
-				// Model supports vision - all models use URLs (no base64 needed)
-				content.push({
-					type: "image" as const,
-					image: file.url,
-				});
-			}
-		}
-		// Handle PDFs
-		else if (file.fileType === "application/pdf") {
-			if (hasPdfSupport && provider === "anthropic") {
-				// Claude supports PDFs as file type
-				content.push({
-					type: "file" as const,
-					data: file.url,
-					mediaType: "application/pdf",
-				});
-			} else {
-				// PDF not supported - add as text description
-				const description = `\n[Attached PDF: ${file.fileName} (${(file.fileSize / 1024).toFixed(1)}KB)] - Note: PDF content analysis requires Claude models.`;
-				content.push({
-					type: "text" as const,
-					text: description,
-				});
-			}
-		}
-		// For other file types, add as text description
-		else {
-			const description = `\n[Attached file: ${file.fileName} (${file.fileType}, ${(file.fileSize / 1024).toFixed(1)}KB)]`;
-
-			if (content[0] && "text" in content[0]) {
-				content[0].text += description;
-			}
-		}
-	}
-
-	return content;
-}
 
 /**
  * Create system prompt based on model capabilities
