+++ apps/www/convex/messages/helpers.ts
@@ -1,12 +1,20 @@
+import type { CoreMessage } from "ai";
+import { stepCountIs, streamText } from "ai";
 import type { Infer } from "convex/values";
+import type { ModelId } from "../../src/lib/ai/schemas.js";
+import {
+	getProviderFromModelId,
+	isThinkingMode,
+} from "../../src/lib/ai/schemas.js";
 import { internal } from "../_generated/api.js";
 import type { Id } from "../_generated/dataModel.js";
 import type { ActionCtx, MutationCtx } from "../_generated/server.js";
-import { clearGenerationFlag as clearGenerationFlagUtil } from "../lib/message_service.js";
+import { createAIClient } from "../lib/ai_client.js";
 import type {
 	modelIdValidator,
 	modelProviderValidator,
 } from "../validators.js";
+import { createWebSearchTool } from "./tools.js";
 import type { MessageUsageUpdate } from "./types.js";
 
 /**
@@ -208,4 +216,167 @@ export async function handleAIResponseError(
 			// This is a critical error that could leave the thread in a locked state
 		}
 	}
+}
+
+/**
+ * Create initial streaming message (moved from lib/message_service.ts)
+ */
+export async function createStreamingMessageUtil(
+	ctx: ActionCtx,
+	threadId: Id<"threads">,
+	modelId: ModelId,
+	streamId: string,
+	usedUserApiKey: boolean,
+): Promise<Id<"messages">> {
+	const provider = getProviderFromModelId(modelId);
+
+	return await ctx.runMutation(internal.messages.createStreamingMessage, {
+		threadId,
+		streamId,
+		provider,
+		modelId,
+		usedUserApiKey,
+	});
+}
+
+/**
+ * Update thread usage with retry logic (moved from lib/message_service.ts)
+ */
+export async function updateThreadUsageUtil(
+	ctx: ActionCtx,
+	threadId: Id<"threads">,
+	modelId: ModelId,
+	usage:
+		| {
+				promptTokens?: number;
+				inputTokens?: number;
+				completionTokens?: number;
+				outputTokens?: number;
+				totalTokens?: number;
+				completionTokensDetails?: { reasoningTokens?: number };
+				reasoningTokens?: number;
+				promptTokensDetails?: { cachedTokens?: number };
+				cachedInputTokens?: number;
+		  }
+		| null
+		| undefined,
+) {
+	if (!usage) return;
+
+	const promptTokens = usage.promptTokens || usage.inputTokens || 0;
+	const completionTokens = usage.completionTokens || usage.outputTokens || 0;
+	const totalTokens = usage.totalTokens || promptTokens + completionTokens;
+
+	await ctx.runMutation(internal.messages.updateThreadUsageMutation, {
+		threadId,
+		usage: {
+			promptTokens,
+			completionTokens,
+			totalTokens,
+			reasoningTokens:
+				usage.completionTokensDetails?.reasoningTokens ||
+				usage.reasoningTokens ||
+				0,
+			cachedTokens:
+				usage.promptTokensDetails?.cachedTokens || usage.cachedInputTokens || 0,
+			modelId,
+		},
+	});
+}
+
+/**
+ * Clear generation flag (moved from lib/message_service.ts)
+ */
+export async function clearGenerationFlagUtil(
+	ctx: ActionCtx,
+	threadId: Id<"threads">,
+) {
+	await ctx.runMutation(internal.messages.clearGenerationFlag, {
+		threadId,
+	});
+}
+
+/**
+ * Stream AI response (moved from lib/message_service.ts)
+ */
+export async function streamAIResponse(
+	ctx: ActionCtx,
+	modelId: ModelId,
+	messages: CoreMessage[],
+	messageId: Id<"messages">,
+	userApiKeys: {
+		openai?: string;
+		anthropic?: string;
+		openrouter?: string;
+	} | null,
+	webSearchEnabled?: boolean,
+) {
+	const provider = getProviderFromModelId(modelId);
+	const aiClient = createAIClient(modelId, userApiKeys);
+
+	// Prepare generation options
+	const generationOptions: Parameters<typeof streamText>[0] = {
+		model: aiClient,
+		messages,
+		temperature: 0.7,
+	};
+
+	// Add web search tool if enabled
+	if (webSearchEnabled) {
+		generationOptions.tools = {
+			web_search: createWebSearchTool(),
+		};
+		generationOptions.stopWhen = stepCountIs(5);
+	}
+
+	// For Claude 4.0 thinking mode, enable thinking/reasoning
+	if (provider === "anthropic" && isThinkingMode(modelId)) {
+		generationOptions.providerOptions = {
+			anthropic: {
+				thinking: {
+					type: "enabled",
+					budgetTokens: 12000,
+				},
+			},
+		};
+	}
+
+	// Use the AI SDK v5 streamText
+	const result = streamText(generationOptions);
+
+	let fullText = "";
+	let hasContent = false;
+	let toolCallsInProgress = 0;
+
+	// Process the stream
+	for await (const chunk of result.textStream) {
+		if (chunk) {
+			fullText += chunk;
+			hasContent = true;
+
+			await ctx.runMutation(internal.messages.addTextPart, {
+				messageId,
+				text: chunk,
+			});
+		}
+	}
+
+	// Process tool calls if web search is enabled
+	if (webSearchEnabled) {
+		for await (const streamPart of result.fullStream) {
+			if (streamPart.type === "tool-call") {
+				toolCallsInProgress++;
+			}
+		}
+	}
+
+	// Get final usage with optional chaining
+	const finalUsage = await result.usage;
+
+	return {
+		fullText,
+		hasContent,
+		toolCallsInProgress,
+		usage: finalUsage,
+	};
 }
